Addition of multiple measurements between transmission
I have managed to make a program with emotions. When it's in a good mood, it collects and transmits the data as required. When it's not feeling so great, it'll transmit the data once before starting to break. Sometimes it hates me and runs through the setup() loop of the Arduino repeatedly. Hmm.